NCA-September-2024/Web/My-Name-Is-Marrie-Biscuit-V2/requirements.txt
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
PyJWT==2.9.0
Werkzeug==3.0.4
gunicorn==23.0.0
